api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-genai==1.26.0
supabase
python-multipart==0.0.6 